+++ Samples~/SimpleRelay/SimpleRelay.cs
@@ -1,10 +1,11 @@
-﻿using System;
+using System;
 using Unity.Services.Relay;
 using Unity.Services.Relay.Models;
 using UnityEngine;
 using UnityEngine.UI;
 using Unity.Services.Core;
 using Unity.Services.Authentication;
+using System.Collections.Generic;
 
 /// <summary>
 /// A simple sample showing how to use the Relay Allocation package. As the host, you can authenticate, request a relay allocation, get a join code and join the allocation.
@@ -19,6 +20,11 @@ public class SimpleRelay : MonoBehaviour
     /// The textbox displaying the Player Id.
     /// </summary>
     public Text PlayerIdText;
+
+    /// <summary>
+    /// The dropdown displaying the region.
+    /// </summary>
+    public Dropdown RegionsDropdown;
 
     /// <summary>
     /// The textbox displaying the Allocation Id.
@@ -39,6 +45,8 @@ public class SimpleRelay : MonoBehaviour
     private Guid playerAllocationId;
     private string joinCode = "n/a";
     private string playerId = "Not signed in";
+    private List<Region> regions = new List<Region>();
+    private List<string> regionOptions = new List<string>();
 
     async void Start()
     {
@@ -50,6 +58,9 @@ public class SimpleRelay : MonoBehaviour
     void UpdateUI()
     {
         PlayerIdText.text = playerId;
+        RegionsDropdown.interactable = regions.Count > 0;
+        RegionsDropdown.options?.Clear();
+        RegionsDropdown.AddOptions(regionOptions);
         HostAllocationIdText.text = hostAllocationId.ToString();
         JoinCodeText.text = joinCode;
         PlayerAllocationIdText.text = playerAllocationId.ToString();
@@ -69,6 +80,23 @@ public class SimpleRelay : MonoBehaviour
     }
 
     /// <summary>
+    /// Event handler for when the Get Regions button is clicked.
+    /// </summary>
+    public async void OnRegion()
+    {
+        Debug.Log("Host - Getting regions.");
+        var allRegions = await Relay.Instance.ListRegionsAsync();
+        regions.Clear();
+        foreach (var region in allRegions)
+        {
+            Debug.Log(region.Id + ": " + region.Description);
+            regionOptions.Add(region.Id);
+            regions.Add(region);
+        }
+        UpdateUI();
+    }
+
+    /// <summary>
     /// Event handler for when the Allocate button is clicked.
     /// </summary>
     public async void OnAllocate()
@@ -76,7 +104,7 @@ public class SimpleRelay : MonoBehaviour
         Debug.Log("Host - Creating an allocation.");
 
         // Important: Once the allocation is created, you have ten seconds to BIND
-        Allocation allocation = await Relay.Instance.CreateAllocationAsync(4);
+        Allocation allocation = await Relay.Instance.CreateAllocationAsync(4, regions.Count > 0 ? regions[RegionsDropdown.value].Id : null);
         hostAllocationId = allocation.AllocationId;
 
         Debug.Log("Host Allocation ID: " + hostAllocationId.ToString());
@@ -124,5 +152,4 @@ public class SimpleRelay : MonoBehaviour
 
         UpdateUI();
     }
-
 }